Split AST type name helpers
